// src/common/types/generated/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IMicrocopyFields {
  /** key */
  key?: string | undefined;

  /** value */
  value?: string | undefined;
}

/** A key-single text value pair */

export interface IMicrocopy extends Entry<IMicrocopyFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "microcopy";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IMicrocopyPageFields {
  /** slug */
  slug?: string | undefined;

  /** microcopy */
  microcopy: (IMicrocopy | IMicrocopyRichText)[];
}

/** An individual page in the application */

export interface IMicrocopyPage extends Entry<IMicrocopyPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "microcopyPage";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IMicrocopyRichTextFields {
  /** key */
  key?: string | undefined;

  /** value */
  value?: Document | undefined;
}

/** A key-rich text value pair */

export interface IMicrocopyRichText extends Entry<IMicrocopyRichTextFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "microcopyRichText";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "microcopy" | "microcopyPage" | "microcopyRichText";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";